+++ app.json
@@ -22,6 +22,9 @@
     ],
     "ios": {
       "supportsTablet": true
+    },
+    "android": {
+      "package": "expo.android.package"
     }
   }
 }
